refactor(lint-cli): split run.ts into planning, sizing and execution
run.ts held three unrelated modules behind nine exports: the ordering-critical
plan phase, the pass sizer, and the child runner. Each is now its own module
with an interface that hides what callers never needed.

sizePasses() absorbs sizePass, the three dirty-count helpers and the ten-field
SizePassContext. The mutate/read-only routing, the cleared-cache short circuit,
the conservative outside-cwd sizing and the typed-pass auto-skip are all
implementation now — callers get planned passes.

execute() folds runConcurrent, runSequential and spawnChild behind one call;
runLint's "fix or single child" branch became its `sequential` argument.

run.ts is 62 lines and exports runLint.
